glGetProgramResourceiv参数使用错误

+++ lume/LumeRender/api/render/gles/intf_device_gles.h
@@ -156,7 +156,7 @@ struct DevicePropertiesGLES {
     int32_t shaderStorageBufferOffsetAlignment;
 
     // ES32
-    int32_t minSampleShadingValue;
+    float  minSampleShadingValue;
     int32_t maxDebugGroupStackDepth;
     int32_t maxDebugLoggedMessages;
     int32_t maxDebugMessageLength;
